+++ apps/rra-prototype/behaviours/check-device-type.js
