+++ apps/rra-prototype/behaviours/clear-session.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = SuperClass => class extends SuperClass {
+  getValues(req, res, next) {
+    req.sessionModel.reset();
+    this.successHandler(req, res, next);
+  }
+};
