+++ apps/rra-prototype/behaviours/remove-image.js
