+++ apps/rra-prototype/behaviours/skip-step.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = superclass => class Skippable extends superclass {
+  get(req, res, next) {
+    this.emit('complete', req, res);
+    super.get(req, res, next);
+  }
+};
