+++ apps/rra-prototype/lib/staticRraGrades
