+++ apps/rra-prototype/lib/staticRraGrouping.js
