+++ apps/rra-prototype/lib/staticRraLevels.js
