+++ apps/rra-prototype/lib/staticRraPortfolio.js
