+++ apps/rra-prototype/lib/staticRraScores.js
