+++ apps/rra-prototype/lib/staticSfiaSkills.js
