+++ apps/rra-prototype/translations/src/en/buttons.json
@@ -1,0 +1,10 @@
+{
+    "continue": "Continue",
+    "login": "Login",
+    "start": "Start Application",
+    "close": "Close",
+    "confirm": "Confirm submission",
+    "start-again": "Start again",
+    "try-again": "Try again",
+    "back": "Back"
+  }
