+++ apps/rra-prototype/translations/src/en/journey.json
@@ -1,0 +1,3 @@
+{
+  "header": "Recruitment and Retention Allowance"
+}
