+++ apps/rra-prototype/views/confirmation.html
@@ -1,0 +1,15 @@
+<title>{{#t}}pages.confirmation.title{{/t}} – GOV.UK</title>
+
+{{<partials-page}}
+  {{$page-content}}
+    <div class="alert-complete" role="alert">
+      <span class="icon icon-complete" role="presentation"></span>
+      <div class="alert-message">
+        <h1>{{#t}}pages.confirmation.alert{{/t}}</h1>
+      </div>
+    </div>
+    <h2>{{#t}}pages.confirmation.subheader{{/t}}</h2>
+    <p>{{#t}}pages.confirmation.content{{/t}}</p>
+    <a href="/" class="button" role="button">{{#t}}buttons.start-again{{/t}}</a>
+  {{/page-content}}
+{{/partials-page}}
