+++ apps/rra-prototype/views/layout.html
@@ -1,0 +1,81 @@
+{{<govuk-template}}
+  {{$head}}
+    {{> partials-head}}
+  {{/head}}
+  {{$pageTitle}}
+    {{#errorlist.length}}{{#t}}errorlist.prefix{{/t}}{{/errorlist.length}}{{$header}}{{/header}} &ndash; GOV.UK
+  {{/pageTitle}}
+  {{$headerClass}}with-proposition{{/headerClass}}
+  {{$insideHeader}}{{/insideHeader}}
+  {{$main}}
+    <main class="main" role="main">
+      <div class="phase-banner">
+        <p>
+          <strong class="phase-tag">{{#t}}journey.phase{{/t}}</strong>
+          <span>This is a new service – your <a href="{{feedbackUrl}}{{^feedbackUrl}}https://eforms.homeoffice.gov.uk/outreach/feedback.ofml{{/feedbackUrl}}">feedback</a> will help us to improve it.</span>
+        </p>
+      </div>
+      <nav class="app-navigation govuk-clearfix">
+        <ul class="app-navigation__list app-width-container">
+
+          <li class="app-navigation__list-item">
+            <a class="govuk-link govuk-link--no-visited-state govuk-link--no-underline app-navigation__link" href="/" data-topnav="Get started">Get started</a>
+          </li>
+
+          <li class="app-navigation__list-item">
+            <a class="govuk-link govuk-link--no-visited-state govuk-link--no-underline app-navigation__link" href="/name?type=basic" data-topnav="Basic form">Basic form</a>
+          </li>
+
+          <li class="app-navigation__list-item">
+            <a class="govuk-link govuk-link--no-visited-state govuk-link--no-underline app-navigation__link" href="/name?type=complex" data-topnav="Complex form">Complex form</a>
+          </li>
+
+          <li class="app-navigation__list-item">
+            <a class="govuk-link govuk-link--no-visited-state govuk-link--no-underline app-navigation__link" href="/build-your-own-form/" data-topnav="Build your own form">Build your own form</a>
+          </li>
+          <li class="app-navigation__list-item">
+            <a class="govuk-link govuk-link--no-visited-state govuk-link--no-underline app-navigation__link" href="/rra-prototype/" data-topnav="RRA Prototype">RRA Prototype</a>
+          </li>
+        </ul>
+      </nav>
+      <div id="post-nav-content">
+        <div>
+          <span id="step">
+            {{> partials-back}} <span>{{$step}}{{/step}}</span>
+          </span>
+        </div>
+        <div class="grid-row" id="content">
+          {{> partials-maincontent-left}}
+        </div>
+        {{> partials-continue}}
+      </div>
+      <script type="text/javascript" src="{{assetPath}}/js/bundle.js"></script>
+    </main>
+  {{/main}}
+  {{$cookieMessage}}
+    {{#gaTagId}}
+      {{> partials-cookie-banner}}
+    {{/gaTagId}}
+    {{^gaTagId}}
+      <p class="no-ga-tag">
+        {{#t}}cookies.banner.message{{/t}}
+        <a href="/cookies">{{#t}}cookies.banner.link{{/t}}</a>
+      </p>
+    {{/gaTagId}}
+  {{/cookieMessage}}
+  {{$footerSupportLinks}}
+    <ul>
+      {{#footerSupportLinks}}
+        <li><a href="{{path}}">{{#t}}{{property}}{{/t}}</a></li>
+      {{/footerSupportLinks}}
+      {{^footerSupportLinks}}
+        <li><a href="/cookies">{{#t}}base.cookies{{/t}}</a></li>
+        <li><a href="/terms-and-conditions">{{#t}}base.terms{{/t}}</a></li>
+        <li><a href="/accessibility">{{#t}}base.accessibility{{/t}}</a></li>
+      {{/footerSupportLinks}}
+    </ul>
+  {{/footerSupportLinks}}
+  {{$bodyEnd}}
+    {{> partials-gatag}}
+  {{/bodyEnd}}
+{{/govuk-template}}
