+++ apps/rra-prototype/views/partials/summary-table.html
@@ -1,0 +1,14 @@
+<div>
+  <h2>{{section}}</h2>
+</div>
+<table class="table-details">
+  <tbody>
+  {{#fields}}
+  <tr>
+    <td>{{label}}</td>
+    <td data-value="{{value}}">{{value}}</td>
+    <td><span><a href="{{baseUrl}}{{step}}/edit#{{field}}" class="link" id="{{field}}-change">{{#t}}buttons.change{{/t}}</a></span></td>
+  </tr>
+  {{/fields}}
+  </tbody>
+</table>
