+++ test/_unit/example_app/remove-image.spec.js
@@ -1,9 +1,9 @@
 'use strict';
 
 const { expect } = require('chai');
-const Behaviour = require('../../../apps/demo/behaviours/remove-image');
+const Behaviour = require('../../../apps/rra-prototype/behaviours/remove-image');
 
-describe("apps/demo/rra 'remove-image' behaviour should ", () => {
+describe("apps/rra-prototype 'remove-image' behaviour should ", () => {
   it('export a function', () => {
     expect(Behaviour).to.be.a('function');
   });
